refactor(client): extract useBlipLlmAssist & useInteractionsLog hooks
Move the logic-heavy query/mutation/derivation code out of BlipLlmAssist and
ResourceInteractionsLog into bundled hooks (mirroring useResourceModules),
leaving the components presentational and under the import/max-dependencies
threshold. UI-close side effects move to per-mutate onSuccess options.

Behavior and public props unchanged.

Fixes #310.
